+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Link multiple authentication types to one account [#1233](https://github.com/rokwire/illinois-app/issues/1233).
+- Canvas Calendar sample UI (in progress) [#1274](https://github.com/rokwire/illinois-app/issues/1274).
 ### Changed
 - Updated Gies Widgets: remove scroll from sub pages [#1291](https://github.com/rokwire/illinois-app/issues/1291).
 - GroupDetailPanel: Allow only Admin to create Poll [#1280](https://github.com/rokwire/illinois-app/issues/1280).
 - Events service moved to Rokwire plugin [#1203](https://github.com/rokwire/illinois-app/issues/1203).
 - Groups service moved to Rokwire plugin [#1203](https://github.com/rokwire/illinois-app/issues/1203).
+### Fixed
+- Loading Canvas Syllabus html content [#1274](https://github.com/rokwire/illinois-app/issues/1274).
 
 ## [3.2.10] - 2022-02-01
 ### Added